backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery[redis]==5.4.0
aiofiles==24.1.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1
